fix: filter non-sandbox stacks
