+++ .changeset/rotten-swans-study.md
@@ -1,0 +1,5 @@
+---
+'@aws-amplify/deployed-backend-client': patch
+---
+
+filter non-sandbox stacks
